feat(election): support deposit reward to election reward
