+++ core/vm/election/electiondb_test.go
@@ -55,8 +55,8 @@ var (
 		StakeCount: big.NewInt(230),
 		TimeStamp:  big.NewInt(1531454152),
 	}
-	bounty = Bounty{
-		RestTotalBounty: big.NewInt(1e18),
+	bounty = Reward{
+		Rest: big.NewInt(1e18),
 	}
 )
 
@@ -141,7 +141,7 @@ func TestConvertToKV(t *testing.T) {
 		t.Error(err)
 	}
 
-	err = convertToKV(BOUNTYPREFIX, bounty, print)
+	err = convertToKV(REWARDPREFIX, bounty, print)
 	if err != nil {
 		t.Error(err)
 	}
@@ -213,13 +213,13 @@ func TestConvertToStruct(t *testing.T) {
 		t.Errorf(err.Error())
 	}
 
-	var bounty1 Bounty
-	err = convertToStruct(BOUNTYPREFIX, contractAddr, &bounty1, getFn)
+	var bounty1 Reward
+	err = convertToStruct(REWARDPREFIX, contractAddr, &bounty1, getFn)
 	if err != nil {
 		t.Error(err)
 	}
-	if bounty1.RestTotalBounty.Cmp(bounty.RestTotalBounty) != 0 {
-		t.Errorf("Error: the reset total Bounty before is %v after is %v", bounty.RestTotalBounty, bounty1.RestTotalBounty)
+	if bounty1.Rest.Cmp(bounty.Rest) != 0 {
+		t.Errorf("Error: the reset total Reward before is %v after is %v", bounty.Rest, bounty1.Rest)
 	}
 }
 
@@ -245,7 +245,7 @@ func TestSetToDB(t *testing.T) {
 		t.Error(err)
 	}
 
-	err = setRestBounty(stateDB, bounty)
+	err = setReward(stateDB, bounty)
 	if err != nil {
 		t.Error(err)
 	}
@@ -262,7 +262,7 @@ func TestGetFromDB(t *testing.T) {
 	err1 := c.setVoter(voter)
 	err2 := c.setCandidate(candidate)
 	err3 := c.setStake(stake)
-	err4 := setRestBounty(stateDB, bounty)
+	err4 := setReward(stateDB, bounty)
 
 	if err1 != nil || err2 != nil || err3 != nil || err4 != nil {
 		t.Fatal("SetToDB err", err1, err2, err3, err4)
@@ -284,9 +284,9 @@ func TestGetFromDB(t *testing.T) {
 		t.Errorf(err.Error())
 	}
 
-	bounty1 := getRestBounty(stateDB)
-	if bounty.RestTotalBounty.Cmp(bounty1.RestTotalBounty) != 0 {
-		t.Errorf("Error: the reset total Bounty before is %v after is %v", bounty.RestTotalBounty, bounty1.RestTotalBounty)
+	bounty1 := getReward(stateDB)
+	if bounty.Rest.Cmp(bounty1.Rest) != 0 {
+		t.Errorf("Error: the reset total Reward before is %v after is %v", bounty.Rest, bounty1.Rest)
 	}
 }
